[web][refactor] Split html and canvaskit classes more (#14320)
* Split surface and canvaskit classes more

* fix licenses

* address comments
