+++ ci/licenses_golden/licenses_flutter
@@ -398,6 +398,7 @@ FILE: ../../../flutter/lib/web_ui/lib/src/engine/compositor/layer.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/compositor/layer_scene_builder.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/compositor/layer_tree.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/compositor/n_way_canvas.dart
+FILE: ../../../flutter/lib/web_ui/lib/src/engine/compositor/painting.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/compositor/path.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/compositor/path_metrics.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/compositor/picture.dart
@@ -426,7 +427,6 @@ FILE: ../../../flutter/lib/web_ui/lib/src/engine/platform_views.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/plugins.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/pointer_binding.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/pointer_converter.dart
-FILE: ../../../flutter/lib/web_ui/lib/src/engine/recording_canvas.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/render_vertices.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/rrect_renderer.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/semantics/accessibility.dart
@@ -451,8 +451,11 @@ FILE: ../../../flutter/lib/web_ui/lib/src/engine/surface/clip.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/surface/debug_canvas_reuse_overlay.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/surface/offset.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/surface/opacity.dart
+FILE: ../../../flutter/lib/web_ui/lib/src/engine/surface/painting.dart
+FILE: ../../../flutter/lib/web_ui/lib/src/engine/surface/path_metrics.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/surface/picture.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/surface/platform_view.dart
+FILE: ../../../flutter/lib/web_ui/lib/src/engine/surface/recording_canvas.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/surface/scene.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/surface/scene_builder.dart
 FILE: ../../../flutter/lib/web_ui/lib/src/engine/surface/surface.dart
