+++ lib/web_ui/lib/src/engine.dart
@@ -38,6 +38,7 @@ part 'engine/compositor/layer_scene_builder.dart';
 part 'engine/compositor/layer_tree.dart';
 part 'engine/compositor/n_way_canvas.dart';
 part 'engine/compositor/path.dart';
+part 'engine/compositor/painting.dart';
 part 'engine/compositor/path_metrics.dart';
 part 'engine/compositor/picture.dart';
 part 'engine/compositor/picture_recorder.dart';
@@ -65,7 +66,6 @@ part 'engine/platform_views.dart';
 part 'engine/plugins.dart';
 part 'engine/pointer_binding.dart';
 part 'engine/pointer_converter.dart';
-part 'engine/recording_canvas.dart';
 part 'engine/render_vertices.dart';
 part 'engine/rrect_renderer.dart';
 part 'engine/semantics/accessibility.dart';
@@ -90,8 +90,11 @@ part 'engine/surface/clip.dart';
 part 'engine/surface/debug_canvas_reuse_overlay.dart';
 part 'engine/surface/offset.dart';
 part 'engine/surface/opacity.dart';
+part 'engine/surface/painting.dart';
+part 'engine/surface/path_metrics.dart';
 part 'engine/surface/picture.dart';
 part 'engine/surface/platform_view.dart';
+part 'engine/surface/recording_canvas.dart';
 part 'engine/surface/scene.dart';
 part 'engine/surface/scene_builder.dart';
 part 'engine/surface/surface.dart';
