+++ lib/web_ui/lib/src/engine/compositor/util.dart
@@ -150,83 +150,6 @@ js.JsObject makeSkBlendMode(ui.BlendMode blendMode) {
   }
 }
 
-js.JsObject makeSkPaint(ui.Paint paint) {
-  if (paint == null) return null;
-
-  final dynamic skPaint = js.JsObject(canvasKit['SkPaint']);
-
-  if (paint.shader != null) {
-    final EngineGradient engineShader = paint.shader;
-    skPaint.callMethod(
-        'setShader', <js.JsObject>[engineShader.createSkiaShader()]);
-  }
-
-  if (paint.color != null) {
-    skPaint.callMethod('setColor', <int>[paint.color.value]);
-  }
-
-  js.JsObject skPaintStyle;
-  switch (paint.style) {
-    case ui.PaintingStyle.stroke:
-      skPaintStyle = canvasKit['PaintStyle']['Stroke'];
-      break;
-    case ui.PaintingStyle.fill:
-      skPaintStyle = canvasKit['PaintStyle']['Fill'];
-      break;
-  }
-  skPaint.callMethod('setStyle', <js.JsObject>[skPaintStyle]);
-
-  js.JsObject skBlendMode = makeSkBlendMode(paint.blendMode);
-  if (skBlendMode != null) {
-    skPaint.callMethod('setBlendMode', <js.JsObject>[skBlendMode]);
-  }
-
-  skPaint.callMethod('setAntiAlias', <bool>[paint.isAntiAlias]);
-
-  if (paint.strokeWidth != 0.0) {
-    skPaint.callMethod('setStrokeWidth', <double>[paint.strokeWidth]);
-  }
-
-  if (paint.maskFilter != null) {
-    final ui.BlurStyle blurStyle = paint.maskFilter.webOnlyBlurStyle;
-    final double sigma = paint.maskFilter.webOnlySigma;
-
-    js.JsObject skBlurStyle;
-    switch (blurStyle) {
-      case ui.BlurStyle.normal:
-        skBlurStyle = canvasKit['BlurStyle']['Normal'];
-        break;
-      case ui.BlurStyle.solid:
-        skBlurStyle = canvasKit['BlurStyle']['Solid'];
-        break;
-      case ui.BlurStyle.outer:
-        skBlurStyle = canvasKit['BlurStyle']['Outer'];
-        break;
-      case ui.BlurStyle.inner:
-        skBlurStyle = canvasKit['BlurStyle']['Inner'];
-        break;
-    }
-
-    final js.JsObject skMaskFilter = canvasKit
-        .callMethod('MakeBlurMaskFilter', <dynamic>[skBlurStyle, sigma, true]);
-    skPaint.callMethod('setMaskFilter', <js.JsObject>[skMaskFilter]);
-  }
-
-  if (paint.imageFilter != null) {
-    final SkImageFilter skImageFilter = paint.imageFilter;
-    skPaint.callMethod(
-        'setImageFilter', <js.JsObject>[skImageFilter.skImageFilter]);
-  }
-
-  if (paint.colorFilter != null) {
-    EngineColorFilter engineFilter = paint.colorFilter;
-    SkColorFilter skFilter = engineFilter._toSkColorFilter();
-    skPaint.callMethod('setColorFilter', <js.JsObject>[skFilter.skColorFilter]);
-  }
-
-  return skPaint;
-}
-
 // Mappings from SkMatrix-index to input-index.
 const List<int> _skMatrixIndexToMatrix4Index = <int>[
   0, 4, 12, // Row 1
