+++ lib/web_ui/test/golden_tests/engine/canvas_rect_golden_test.dart
@@ -36,27 +36,27 @@ void main() async {
 void paintRects(BitmapCanvas canvas) {
 
     canvas.drawRect(Rect.fromLTRB(30, 40, 100, 50),
-      PaintData()
+      SurfacePaintData()
         ..color = Color(0xFF4CAF50) //Colors.green
         ..strokeWidth = 1.0
         ..style = PaintingStyle.stroke);
 
     // swap left and right.
     canvas.drawRect(Rect.fromLTRB(100, 150, 30, 140),
-      PaintData()
+      SurfacePaintData()
         ..color = Color(0xFFF44336) //Colors.red
         ..strokeWidth = 1.0
         ..style = PaintingStyle.stroke);
 
     // Repeat above for fill
     canvas.drawRect(Rect.fromLTRB(30, 240, 100, 250),
-      PaintData()
+      SurfacePaintData()
         ..color = Color(0xFF4CAF50) //Colors.green
         ..style = PaintingStyle.fill);
 
     // swap left and right.
     canvas.drawRect(Rect.fromLTRB(100, 350, 30, 340),
-      PaintData()
+      SurfacePaintData()
         ..color = Color(0xFFF44336) //Colors.red
         ..style = PaintingStyle.fill);
 }
