+++ lib/web_ui/test/mock_engine_canvas.dart
@@ -121,7 +121,7 @@ class MockEngineCanvas implements EngineCanvas {
   }
 
   @override
-  void drawLine(Offset p1, Offset p2, PaintData paint) {
+  void drawLine(Offset p1, Offset p2, SurfacePaintData paint) {
     _called('drawLine', arguments: <String, dynamic>{
       'p1': p1,
       'p2': p2,
@@ -130,17 +130,17 @@ class MockEngineCanvas implements EngineCanvas {
   }
 
   @override
-  void drawPaint(PaintData paint) {
+  void drawPaint(SurfacePaintData paint) {
     _called('drawPaint', arguments: paint);
   }
 
   @override
-  void drawRect(Rect rect, PaintData paint) {
+  void drawRect(Rect rect, SurfacePaintData paint) {
     _called('drawRect', arguments: paint);
   }
 
   @override
-  void drawRRect(RRect rrect, PaintData paint) {
+  void drawRRect(RRect rrect, SurfacePaintData paint) {
     _called('drawRRect', arguments: <String, dynamic>{
       'rrect': rrect,
       'paint': paint,
@@ -148,7 +148,7 @@ class MockEngineCanvas implements EngineCanvas {
   }
 
   @override
-  void drawDRRect(RRect outer, RRect inner, PaintData paint) {
+  void drawDRRect(RRect outer, RRect inner, SurfacePaintData paint) {
     _called('drawDRRect', arguments: <String, dynamic>{
       'outer': outer,
       'inner': inner,
@@ -157,7 +157,7 @@ class MockEngineCanvas implements EngineCanvas {
   }
 
   @override
-  void drawOval(Rect rect, PaintData paint) {
+  void drawOval(Rect rect, SurfacePaintData paint) {
     _called('drawOval', arguments: <String, dynamic>{
       'rect': rect,
       'paint': paint,
@@ -165,7 +165,7 @@ class MockEngineCanvas implements EngineCanvas {
   }
 
   @override
-  void drawCircle(Offset c, double radius, PaintData paint) {
+  void drawCircle(Offset c, double radius, SurfacePaintData paint) {
     _called('drawCircle', arguments: <String, dynamic>{
       'c': c,
       'radius': radius,
@@ -174,7 +174,7 @@ class MockEngineCanvas implements EngineCanvas {
   }
 
   @override
-  void drawPath(Path path, PaintData paint) {
+  void drawPath(Path path, SurfacePaintData paint) {
     _called('drawPath', arguments: <String, dynamic>{
       'path': path,
       'paint': paint,
@@ -193,7 +193,7 @@ class MockEngineCanvas implements EngineCanvas {
   }
 
   @override
-  void drawImage(Image image, Offset p, PaintData paint) {
+  void drawImage(Image image, Offset p, SurfacePaintData paint) {
     _called('drawImage', arguments: <String, dynamic>{
       'image': image,
       'p': p,
@@ -202,7 +202,7 @@ class MockEngineCanvas implements EngineCanvas {
   }
 
   @override
-  void drawImageRect(Image image, Rect src, Rect dst, PaintData paint) {
+  void drawImageRect(Image image, Rect src, Rect dst, SurfacePaintData paint) {
     _called('drawImageRect', arguments: <String, dynamic>{
       'image': image,
       'src': src,
@@ -221,7 +221,7 @@ class MockEngineCanvas implements EngineCanvas {
 
   @override
   void drawVertices(Vertices vertices, BlendMode blendMode,
-      PaintData paint) {
+      SurfacePaintData paint) {
     _called('drawVertices', arguments: <String, dynamic>{
       'vertices': vertices,
       'blendMode': blendMode,
